Add useFormState hook
